Create README.md
Traversal of Graph using DFS and BFS
In this assignment, you will be implementing Depth-First Search (DFS) and Breadth-First Search (BFS) algorithms to perform traversal on the graph provided in Exercise 2.

Understanding DFS and BFS
Before you begin, it's important to understand the concepts of DFS and BFS.

DFS (Depth-First Search): DFS is an algorithm for traversing or searching tree or graph data structures. It starts at the root node and explores as far as possible along each branch before backtracking.

BFS (Breadth-First Search): BFS is an algorithm for traversing or searching tree or graph data structures. It starts at the tree root (or some arbitrary node of a graph) and explores the neighbor nodes at the present depth prior to moving on to nodes at the next depth level.

Implementing DFS and BFS
DFS Implementation: You will need to write a function or method to perform DFS traversal on the given graph. Make sure to understand the recursive nature of DFS and how it explores as far as possible along each branch before backtracking.

BFS Implementation: Similarly, you will need to write a function or method to perform BFS traversal on the given graph. Understand how BFS explores the neighbor nodes at the present depth before moving on to nodes at the next depth level.